Add AST builder utils to Common Utils
