+++ composer/modules/server/services/ballerina-parser/pom.xml
@@ -47,5 +47,9 @@
             <groupId>com.google.guava</groupId>
             <artifactId>guava</artifactId>
         </dependency>
+        <dependency>
+            <groupId>org.ballerinalang</groupId>
+            <artifactId>language-server-core</artifactId>
+        </dependency>
     </dependencies>
 </project>
